chore: remove ckb-channel dep from ckb-types
