+++ util/types/src/core/mod.rs
@@ -14,7 +14,6 @@
 pub mod cell;
 pub mod error;
 pub mod hardfork;
-pub mod service;
 pub mod tx_pool;
 
 #[cfg(test)]
